Moved login feature to auth folder
