+++ lib/src/core/router/app_router.dart
@@ -1,6 +1,6 @@
 import 'package:auto_route/auto_route.dart';
-import 'package:carey/src/features/login/presentation/views/login_methods_view.dart';
-import 'package:carey/src/features/login/presentation/views/login_view.dart';
+import 'package:carey/src/features/auth/presentation/views/login_methods_view.dart';
+import 'package:carey/src/features/auth/presentation/views/login_view.dart';
 import 'package:carey/src/features/splash/presentation/views/splash_view.dart';
 import 'package:carey/src/features/start/presentation/views/start_view.dart';
 import 'package:carey/src/features/start/presentation/views/welcome_view.dart';
