+++ lib/src/features/auth/data/models/social_login_method.dart
