+++ lib/src/features/auth/domain/entities/login_response_entity.dart
@@ -1,0 +1,7 @@
+import 'package:carey/src/features/auth/data/models/login_response.dart';
+
+class LoginResponseEntity {
+  final LoginData userData;
+
+  LoginResponseEntity(this.userData);
+}
