+++ lib/src/features/auth/presentation/widgets/auth_switcher.dart
