+++ lib/src/features/auth/presentation/widgets/login/social_login_buttons.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+import 'package:carey/src/core/widgets/my_sized_box.dart';
+import 'package:carey/src/features/auth/data/models/social_login_method.dart';
+import 'package:carey/src/features/auth/presentation/widgets/login/login_via_social_button.dart';
+
+class SocialLoginButtons extends StatelessWidget {
+  const SocialLoginButtons({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return ListView.separated(
+      shrinkWrap: true,
+      physics: const NeverScrollableScrollPhysics(),
+      itemBuilder: (_, index) => LoginViaSocialButton(
+        method: SocialLoginMethod.socialLoginMethods[index],
+      ),
+      separatorBuilder: (_, __) => MySizedBox.height17,
+      itemCount: SocialLoginMethod.socialLoginMethods.length,
+    );
+  }
+}
