--- lib/src/features/login/domain/entities/login_response_entity.dart
@@ -1,7 +1,0 @@
-import 'package:carey/src/features/login/data/models/login_response.dart';
-
-class LoginResponseEntity {
-  final LoginData userData;
-
-  LoginResponseEntity(this.userData);
-}
